# Update field to improve performance