I updated the setter within Fields.php

We didn't change to overall logic. But we gained some minor performance improvement while getting a big readability improvement.

@flow-control was kind enough to make some perfomance checks

**Before update**

`
benchValueFromField.......I0 [μ Mo]/r: 8.280 8.280 (μs) [μSD μRSD]/r: 0.000μs 0.00%
 benchRawValueFromField.....I0 [μ Mo]/r: 1.955 1.955 (μs) [μSD μRSD]/r: 0.000μs 0.00%
 benchValueFromRawField.....I0 [μ Mo]/r: 2.026 2.026 (μs) [μSD μRSD]/r: 0.000μs 0.00%
 benchRawValueFromRawField..I0 [μ Mo]/r: 2.564 2.564 (μs) [μSD μRSD]/r: 0.000μs 0.00%
`

**After changes**

`
benchValueFromField.......I0 [μ Mo]/r: 7.714 7.714 (μs) [μSD μRSD]/r: 0.000μs 0.00%
benchRawValueFromField.....I0 [μ Mo]/r: 1.961 1.961 (μs) [μSD μRSD]/r: 0.000μs 0.00%
benchValueFromRawField.....I0 [μ Mo]/r: 2.065 2.065 (μs) [μSD μRSD]/r: 0.000μs 0.00%
benchRawValueFromRawField..I0 [μ Mo]/r: 2.285 2.285 (μs) [μSD μRSD]/r: 0.000μs 0.00%
`

This seems not much, but these functions are called a lot. 

![image](https://user-images.githubusercontent.com/35966734/75794141-abe73c80-5d70-11ea-9241-189ac7e83c8b.png)