@CLAassistant 2020-03-03
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/OXID-eSales/oxideshop_ce?pullRequest=771) <br/>Thank you for your submission! We really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/OXID-eSales/oxideshop_ce?pullRequest=771) before we can accept your contribution.<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/OXID-eSales/oxideshop_ce?pullRequest=771) it.</sub>