@flow-control 2020-03-04
I like it @AlfonsMartin

- it makes the code readable
- and more expressive
- does not change behaviour
- is proven to be faster in 1 of 4 cases
- and performs the same in the other 3 of 4 cases

LGTM!